+++ src/components/layouts/commonLayout.tsx
@@ -1,30 +1,30 @@
 import styled from 'styled-components';
 import React, { useState } from 'react';
 import SideBar from './sideBar';
-import ProfileHeader from './profileHeader';
+import * as Style from '../../css/Layout/Header';
 
 const Container = styled.div`
-margin-left : 12.5vw;
+  margin-left: 12.5vw;
 `;
 
 const LayoutBox = styled.div`
   display: flex;
 `;
 
-
 const Children = styled.div``;
 
-const profileLayout = ({ children }: { children: any }) => {
-
+const CommonLayout = ({ children,title }: { children: any, title : string }) => {
   return (
     <LayoutBox>
       <SideBar />
       <Container>
-        <ProfileHeader />
+        <Style.Header isTeample={false}>
+          <Style.HomeTitle>{title}</Style.HomeTitle>
+        </Style.Header>
         <Children>{children}</Children>
       </Container>
     </LayoutBox>
   );
 };
 
-export default profileLayout;
+export default CommonLayout;
