+++ src/pages/FilePage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import FileHeader from 'components/filePage/FileHeader';
-import Layout from 'components/layouts/layout';
+import Layout from 'components/layouts/teampleLayout';
 import FileList from 'components/filePage/fileList';
 
 const FilePage = () => {
