+++ src/pages/ProfilePage.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import EditProfile from 'components/editProfilePage/editProfile';
-import ProfileLayout from 'components/layouts/profileLayout';
+import ProfileLayout from 'components/layouts/commonLayout';
 
 const ProfilePage = () => {
   return (
-    <ProfileLayout>
+    <ProfileLayout title='프로필'>
       <EditProfile />
     </ProfileLayout>
   );
